RavenDB-25281 extract SpatialVector, Highlighting, Sorting to partial classes
Resolution.cs 5280 → 4136 lines (-22%). Pure extraction, no logic changes.

- QueryPlanBuilder.SpatialVector.cs (712 lines): HandleSpatial, HandleVector,
  ResolveVectorItems, AttachSpatialAndVectorClauses, AttachPostFilterPhases,
  ResolveSpatialFromBindings, ResolveVectorFromBindings, VectorHelpers class.
- QueryPlanBuilder.Highlighting.cs (125 lines): PopulateHighlightingTerms,
  PopulateHighlightingForClause, GetHighlightingValues.
- QueryPlanBuilder.Sorting.cs (376 lines): GetSortMetadata, OrderBy,
  ApplyScoreOrdering, TrySetSortSeekHint, ReplaceAnalyzerForWildcardQueries.
