+++ src/Raven.Server/Documents/Indexes/Persistence/Corax/QueryPlanBuilder.Resolution.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
-using System.IO;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Text;
@@ -15,23 +14,15 @@ using Corax.Querying.Matches.SortingMatches.Meta;
 using Corax.Querying.Planning;
 using Corax.Mappings;
 using Corax.Utils;
-using Raven.Client.Documents.Indexes.Vector;
 using Raven.Client.Exceptions;
-using Raven.Client.Exceptions.Corax;
-using VectorOptions = Raven.Client.Documents.Indexes.Vector.VectorOptions;
-using Raven.Server.Documents.ETL.Providers.AI.Embeddings;
 using Raven.Server.Documents.Indexes.Persistence.Corax.QueryOptimizer;
-using Raven.Server.Documents.Indexes.VectorSearch;
 using Raven.Server.Documents.Queries;
 using Raven.Server.Documents.Queries.AST;
-using Spatial4n.Shapes;
 using Sparrow;
 using Sparrow.Json;
 using Constants = Corax.Constants;
 using RavenConstants = Raven.Client.Constants;
-using SpatialUnits = Raven.Client.Documents.Indexes.Spatial.SpatialUnits;
 using IndexSearcher = Corax.Querying.IndexSearcher;
-using SpatialRelation = Corax.Utils.Spatial.SpatialRelation;
 
 namespace Raven.Server.Documents.Indexes.Persistence.Corax;
 
@@ -44,10 +35,6 @@ namespace Raven.Server.Documents.Indexes.Persistence.Corax;
 /// </summary>
 internal static partial class QueryPlanBuilder
 {
-
-    /// <summary>Maximum number of ORDER BY fields supported by Corax.</summary>
-    private const int MaxSortFields = 16;
-
     /// <summary>
     /// Plan → compile → resolve pipeline. On cache hit (template exists for this query text),
     /// skips AST parsing — re-resolves parameter values from the blittable, re-estimates
@@ -893,55 +880,6 @@ internal static partial class QueryPlanBuilder
         return execList.ToArray();
     }
 
-    /// <summary>Populate ClauseExecution slots for any template-recorded spatial / vector
-    /// post-filter clauses (each a separate phase outside the main bitmap pipeline), then
-    /// call <see cref="AttachPostFilterPhases"/> to wire them onto the plan. No-op when
-    /// the template has neither.</summary>
-    private static void AttachSpatialAndVectorClauses(
-        QueryExecution exec, bool allNegated, PlanTemplate template, QueryPlanBuilder.PlanParameters planParams,
-        QueryBuilderParameters builderParameters, ValueWriter writer)
-    {
-        if (template.SpatialClauses == null && template.VectorClauses == null)
-            return;
-
-        ClauseInfo[] spatialArr = null;
-        ClauseInfo[] vectorArr = null;
-        ClauseExecution[] spatialExecs = null;
-        ClauseExecution[] vectorExecs = null;
-
-        if (template.SpatialClauses != null)
-        {
-            int sLen = template.SpatialClauses.Count;
-            spatialArr = new ClauseInfo[sLen];
-            spatialExecs = new ClauseExecution[sLen];
-            for (int si = 0; si < sLen; si++)
-            {
-                var sc = template.SpatialClauses[si];
-                var scExec = new ClauseExecution(sc);
-                PopulateClauseValues(scExec, planParams.QueryParameters, writer, builderParameters);
-                spatialArr[si] = sc;
-                spatialExecs[si] = scExec;
-            }
-        }
-
-        if (template.VectorClauses != null)
-        {
-            int vLen = template.VectorClauses.Count;
-            vectorArr = new ClauseInfo[vLen];
-            vectorExecs = new ClauseExecution[vLen];
-            for (int vi = 0; vi < vLen; vi++)
-            {
-                var vc = template.VectorClauses[vi];
-                var vcExec = new ClauseExecution(vc);
-                PopulateClauseValues(vcExec, planParams.QueryParameters, writer, builderParameters);
-                vectorArr[vi] = vc;
-                vectorExecs[vi] = vcExec;
-            }
-        }
-
-        AttachPostFilterPhases(exec, allNegated, spatialArr, spatialExecs, vectorArr, vectorExecs);
-    }
-
     /// <summary>Remap template-position optimization indices (sort-driving, compound-exact pair,
     /// compound-field driver) to post-sort runtime indices. After WHEN elimination + cardinality
     /// sort, clauses are reordered relative to the template; each clause carries its
@@ -1291,108 +1229,6 @@ internal static partial class QueryPlanBuilder
         exec.PackedParamValue = new PackedParam(packedType, startIdx);
         exec.InTermCount = nonNullCount;
         exec.HasNullTerm = hasNullTerm;
-    }
-
-    /// <summary>Resolve spatial parameters from cached bindings (no MethodExpression dependency).</summary>
-    private static void ResolveSpatialFromBindings(ClauseExecution exec, BlittableJsonReaderObject queryParameters)
-    {
-        var bindings = exec.Clause.Bindings;
-        var sp = new SpatialParams();
-
-        // [0] = distanceErrorPct
-        if (bindings.Length > 0 && bindings[BindingIndex.SpatialDistErrPct] != null)
-        {
-            var (depVal, _) = ResolveBindingScalar(bindings[BindingIndex.SpatialDistErrPct], queryParameters, builderParameters: null);
-            sp.DistanceErrorPct = depVal != null ? Convert.ToDouble(depVal) : -1;
-        }
-
-        // Shape type determined by the number of bindings:
-        // circle has 5 (distErrPct, radius, lat, lng, units), WKT has 3 (distErrPct, wkt, units)
-        if (bindings.Length >= BindingIndex.SpatialCircleBindingCount - 1) // circle: at least distErrPct + radius + lat + lng
-        {
-            sp.ShapeType = SpatialShapeType.Circle;
-            var (r, _) = ResolveBindingScalar(bindings[BindingIndex.SpatialRadius], queryParameters, builderParameters: null);
-            var (lat, _) = ResolveBindingScalar(bindings[BindingIndex.SpatialLatitude], queryParameters, builderParameters: null);
-            var (lng, _) = ResolveBindingScalar(bindings[BindingIndex.SpatialLongitude], queryParameters, builderParameters: null);
-            sp.CircleRadius = Convert.ToDouble(r);
-            sp.CircleLatitude = Convert.ToDouble(lat);
-            sp.CircleLongitude = Convert.ToDouble(lng);
-            if (bindings.Length > BindingIndex.SpatialUnits && bindings[BindingIndex.SpatialUnits] != null)
-            {
-                var (u, _) = ResolveBindingScalar(bindings[BindingIndex.SpatialUnits], queryParameters, builderParameters: null);
-                if (u != null && Enum.TryParse(typeof(SpatialUnits), u.ToString(), true, out var su))
-                    sp.Units = (SpatialUnits)su == SpatialUnits.Kilometers
-                        ? global::Corax.Utils.Spatial.SpatialUnits.Kilometers
-                        : global::Corax.Utils.Spatial.SpatialUnits.Miles;
-            }
-        }
-        else // WKT: distErrPct, wkt, [units]
-        {
-            sp.ShapeType = SpatialShapeType.Wkt;
-            if (bindings.Length > BindingIndex.SpatialWkt && bindings[BindingIndex.SpatialWkt] != null)
-            {
-                var (wkt, _) = ResolveBindingScalar(bindings[BindingIndex.SpatialWkt], queryParameters, builderParameters: null);
-                sp.Wkt = wkt?.ToString();
-                if (bindings.Length > BindingIndex.SpatialWktUnits && bindings[BindingIndex.SpatialWktUnits] != null)
-                {
-                    var (u, _) = ResolveBindingScalar(bindings[BindingIndex.SpatialWktUnits], queryParameters, builderParameters: null);
-                    if (u != null && Enum.TryParse(typeof(SpatialUnits), u.ToString(), true, out var su))
-                        sp.Units = (SpatialUnits)su == SpatialUnits.Kilometers
-                            ? global::Corax.Utils.Spatial.SpatialUnits.Kilometers
-                            : global::Corax.Utils.Spatial.SpatialUnits.Miles;
-                }
-            }
-        }
-
-        exec.Spatial = sp;
-    }
-
-    /// <summary>Resolve vector parameters from cached bindings (no MethodExpression dependency).</summary>
-    private static void ResolveVectorFromBindings(ClauseExecution exec, BlittableJsonReaderObject queryParameters)
-    {
-        var bindings = exec.Clause.Bindings;
-
-        var vec = exec.Vector = new VectorParams { Method = exec.Clause.VectorMethod };
-
-        // [1]=minimumMatch, [2]=numberOfCandidates, [3]=aiTask
-        if (bindings.Length > BindingIndex.VectorMinMatch && bindings[BindingIndex.VectorMinMatch] != null)
-        {
-            var (simVal, _) = ResolveBindingScalar(bindings[BindingIndex.VectorMinMatch], queryParameters, builderParameters: null);
-            vec.MinimumMatch = simVal switch
-            {
-                double d => (float)d,
-                long l => l,
-                _ => -1
-            };
-        }
-
-        if (bindings.Length > BindingIndex.VectorCandidates && bindings[BindingIndex.VectorCandidates] != null)
-        {
-            var (candVal, candType) = ResolveBindingScalar(bindings[BindingIndex.VectorCandidates], queryParameters, builderParameters: null);
-            if (candType != ParamValueType.Null)
-                vec.NumberOfCandidates = Convert.ToInt32(candVal);
-        }
-
-        if (bindings.Length > BindingIndex.VectorAiTask && bindings[BindingIndex.VectorAiTask] != null)
-        {
-            var (taskVal, _) = ResolveBindingScalar(bindings[BindingIndex.VectorAiTask], queryParameters, builderParameters: null);
-            vec.AiTaskName = taskVal?.ToString();
-        }
-
-        // [0]=vector value (may be scalar, array, or blittable object)
-        if (bindings.Length > BindingIndex.VectorValue && bindings[BindingIndex.VectorValue] != null)
-        {
-            var (val, valType) = ResolveBindingRaw(bindings[BindingIndex.VectorValue], queryParameters);
-            vec.ResolvedValue = val;
-            vec.ResolvedValueType = valType;
-            // For scalar parameters, resolve the native type
-            if (valType == ParamValueType.Parameter && val is not (BlittableJsonReaderArray or BlittableJsonReaderObject))
-            {
-                var (resolved, resolvedType) = ResolveParameterValue(val);
-                vec.ResolvedValue = resolved;
-                vec.ResolvedValueType = ToParamValueType(resolvedType);
-            }
-        }
     }
 
     /// <summary>Look up a binding's value from the blittable. Returns the RAW value —
@@ -2436,595 +2272,6 @@ internal static partial class QueryPlanBuilder
         }
     }
 
-    // ── Highlighting ─────────────────────────────────────────────────────
-
-    /// <summary>
-    /// Populate the highlighting terms dictionary from the plan's clauses.
-    /// The old CoraxQueryBuilder did this as a side effect during query building.
-    /// The bitmap pipeline must do it explicitly after plan building.
-    /// </summary>
-    private static void PopulateHighlightingTerms(QueryExecution exec, Dictionary<string, CoraxHighlightingTermIndex> highlightingTerms, QueryMetadata metadata)
-    {
-        var execs = exec.Executions;
-        if (highlightingTerms == null || execs is not { Length: > 0 })
-            return;
-
-        for (int ci = 0; ci < execs.Length; ci++)
-        {
-            var clauseObj = execs[ci].Clause;
-            var clauseExec2 = execs[ci];
-
-            // Recurse into sub-clauses before checking FieldName: OrGroup/AndGroup have
-            // FieldName==null (they are structural wrappers, not field clauses), so a
-            // FieldName-first guard would skip their children entirely.
-            switch (clauseObj?.ClauseType)
-            {
-                case ClauseType.OrGroup or ClauseType.AndGroup when clauseObj.SubClauses is { Count: > 0 }:
-                {
-                    for (int si = 0; si < clauseObj.SubClauses.Count; si++)
-                    {
-                        var subExec = clauseExec2?.SubExecutions != null && si < clauseExec2.SubExecutions.Length ? clauseExec2.SubExecutions[si] : null;
-                        PopulateHighlightingForClause(clauseObj.SubClauses[si], subExec, highlightingTerms, metadata, exec);
-                    }
-
-                    break;
-                }
-            }
-
-            if (clauseObj?.FieldName == null)
-                continue;
-
-            PopulateHighlightingForClause(clauseObj, clauseExec2, highlightingTerms, metadata, exec);
-        }
-    }
-
-    private static void PopulateHighlightingForClause(ClauseInfo clause, ClauseExecution exec, Dictionary<string, CoraxHighlightingTermIndex> highlightingTerms, QueryMetadata metadata, QueryExecution queryExec)
-    {
-        string fieldName = clause.FieldName;
-        if (fieldName == null)
-            return;
-
-        // Skip highlighting for null-valued clauses (e.g. WHERE City == null).
-        // Null is not a search term — there's nothing to highlight. Without this
-        // guard, the highlighter produces a spurious result for null fields (#4781).
-        if (clause.ClauseType is ClauseType.Equals or ClauseType.NotEquals)
-        {
-            var packed = exec?.PackedParamValue ?? PackedParam.None;
-            if (packed.IsNone || (packed is { ValueType: PackedParam.TypeString, Param1: >= 0 } && queryExec.StringValues != null
-                                                                                                && packed.Param1 < queryExec.StringValues.Length
-                                                                                                && queryExec.StringValues[packed.Param1] == null))
-            {
-                return;
-            }
-        }
-
-        if (highlightingTerms.TryGetValue(fieldName, out var existingTerm))
-        {
-            existingTerm.Values ??= GetHighlightingValues(clause, exec, queryExec);
-            return;
-        }
-
-        var term = new CoraxHighlightingTermIndex
-        {
-            FieldName = fieldName,
-            Values = GetHighlightingValues(clause, exec, queryExec)
-        };
-
-        if (metadata.IsDynamic && clause.ClauseType == ClauseType.Search)
-            term.DynamicFieldName = AutoIndexField.GetSearchAutoIndexFieldName(fieldName);
-        else if (metadata.IsDynamic && clause.IsExact)
-            term.DynamicFieldName = AutoIndexField.GetExactAutoIndexFieldName(fieldName);
-
-        highlightingTerms[fieldName] = term;
-
-        // For dynamic indexes, also add the dynamic field name variant
-        if (term.DynamicFieldName != null)
-            highlightingTerms[term.DynamicFieldName] = term;
-    }
-
-    private static object GetHighlightingValues(ClauseInfo clause, ClauseExecution exec, QueryExecution queryExec)
-    {
-        var packed = exec?.PackedParamValue ?? PackedParam.None;
-        if (clause.ClauseType == ClauseType.Between)
-        {
-            return new Tuple<string, string>(
-                FormatValueFromPlan(packed, queryExec),
-                FormatValue2FromPlan(packed, queryExec));
-        }
-
-        int inTermCount = exec?.InTermCount ?? 0;
-        bool hasNullTerm = exec?.HasNullTerm ?? false;
-        if (clause.ClauseType is ClauseType.In or ClauseType.AllIn && (inTermCount > 0 || hasNullTerm))
-        {
-            var p = packed;
-            var terms = new List<string>(inTermCount + (hasNullTerm ? 1 : 0));
-            for (int t = 0; t < inTermCount; t++)
-                terms.Add(FormatValueFromPlan(new PackedParam(p.ValueType, p.Param1 + t), queryExec));
-            if (hasNullTerm)
-                terms.Add(null);
-            return terms;
-        }
-
-        return FormatValueFromPlan(packed, queryExec);
-    }
-
-    // ── Vector / Spatial resolution ──────────────────────────────────────
-
-    /// <summary>
-    /// Resolve vector select operations from the plan into CoraxVectorItem instances.
-    /// These are NOT materialized yet — the caller materializes them with the bitmap-producing
-    /// match as the filterQuery. Returns null if the plan has no vectors.
-    /// </summary>
-    private static CoraxVectorItem[] ResolveVectorItems(QueryExecution exec, QueryBuilderParameters builderParams)
-    {
-        if (exec.VectorSelects == null || exec.VectorSelects.Length == 0)
-            return null;
-
-        var items = new CoraxVectorItem[exec.VectorSelects.Length];
-        for (int i = 0; i < exec.VectorSelects.Length; i++)
-        {
-            var clause = exec.VectorSelects[i].Clause;
-            var vecExec = exec.VectorSelects[i].Exec;
-            if (clause == null || clause.ClauseType != ClauseType.Vector || builderParams == null)
-                throw new InvalidOperationException("Vector select references an invalid clause at index " + i);
-
-            items[i] = HandleVector(builderParams, clause, vecExec, false);
-        }
-
-        return items;
-    }
-
-    private static IQueryMatch HandleSpatial(QueryBuilderParameters builderParameters, ClauseInfo clause, ClauseExecution exec, SpatialOperationType spatialMethod)
-    {
-        var index = builderParameters.Index;
-        var allocator = builderParameters.Allocator;
-
-        // Field name was pre-resolved during parsing.
-        string fieldName = clause.FieldName
-                           ?? throw new InvalidOperationException("Spatial clause has no pre-resolved field name.");
-
-        var fieldMetadata = QueryBuilderHelper.GetFieldMetadata(allocator, fieldName, index, builderParameters.IndexFieldsMapping,
-            builderParameters.FieldsToFetch, builderParameters.HasDynamics, builderParameters.DynamicFields, hasBoost: builderParameters.HasBoost);
-
-        var sp = exec.Spatial;
-        var distanceErrorPct = sp.DistanceErrorPct >= 0
-            ? sp.DistanceErrorPct
-            : RavenConstants.Documents.Indexing.Spatial.DefaultDistanceErrorPct;
-
-        var spatialField = builderParameters.Factories.GetSpatialFieldFactory(fieldName);
-
-        // Build shape from pre-resolved parameters — no GetValue calls
-        IShape shape;
-        SpatialUnits? units = sp.Units.HasValue ? (SpatialUnits)sp.Units.Value : null;
-        if (sp.ShapeType == SpatialShapeType.Circle)
-        {
-            shape = spatialField.ReadCircle(sp.CircleRadius, sp.CircleLatitude, sp.CircleLongitude, units);
-        }
-        else if (sp.Wkt != null)
-        {
-            shape = spatialField.ReadShape(sp.Wkt, units);
-        }
-        else
-        {
-            throw new InvalidOperationException("Spatial clause has no pre-resolved shape parameters.");
-        }
-
-        return builderParameters.IndexSearcher.SpatialQuery(fieldMetadata, distanceErrorPct, shape, spatialField.GetContext(), (SpatialRelation)spatialMethod, token: builderParameters.Token);
-    }
-
-    private static CoraxVectorItem HandleVector(QueryBuilderParameters builderParameters, ClauseInfo clause, ClauseExecution exec, bool exact)
-    {
-        IndexField indexField;
-        string embeddingsGenerationTaskIdentifier;
-
-        var vec = exec.Vector;
-        var minimumMatch = vec.MinimumMatch >= 0
-            ? vec.MinimumMatch
-            : builderParameters.Index.Configuration.CoraxVectorSearchDefaultMinimumSimilarity;
-
-        int numberOfCandidates = vec.NumberOfCandidates >= 0
-            ? vec.NumberOfCandidates
-            : builderParameters.Index.Configuration.CoraxVectorDefaultNumberOfCandidatesForQuerying;
-
-        var fieldName = clause.FieldName
-                        ?? throw new InvalidOperationException("Vector clause has no pre-resolved field name.");
-
-        var fieldMetadata = QueryBuilderHelper.GetFieldMetadata(builderParameters, fieldName, hasBoost: builderParameters.HasBoost);
-
-        // Use pre-resolved vector value and method kind from parsing
-        object methodParameter = vec.ResolvedValue;
-        ValueTokenType valueTokenType = ToValueTokenType(vec.ResolvedValueType);
-
-        if (vec.Method != VectorSourceKind.Inline)
-        {
-            var method = vec.Method switch
-            {
-                VectorSourceKind.FromDocument => VectorHelpers.MethodVectorValue.ForDocument,
-                VectorSourceKind.FromText => VectorHelpers.MethodVectorValue.EmbeddingText,
-                _ => throw new InvalidDataException($"Unknown vector source kind: {vec.Method}")
-            };
-
-            if (method is not VectorHelpers.MethodVectorValue.EmbeddingText)
-            {
-                return (method, methodParameter) switch
-                {
-                    (method: VectorHelpers.MethodVectorValue.ForDocument, string docId) => CoraxVectorItem.BuildForDocVector(builderParameters, fieldMetadata, docId, numberOfCandidates, minimumMatch, exact),
-                    (method: VectorHelpers.MethodVectorValue.ForDocument, StringSegment docIdSegment) => CoraxVectorItem.BuildForDocVector(builderParameters, fieldMetadata, docIdSegment.Value, numberOfCandidates, minimumMatch, exact),
-                    (method: VectorHelpers.MethodVectorValue.ForRaw, string vectorAsBase64) => CoraxVectorItem.BuildSingleVector(builderParameters, fieldMetadata,
-                        GenerateEmbeddings.FromBase64Array(VectorOptions.Default, builderParameters.Allocator, vectorAsBase64), numberOfCandidates, minimumMatch, exact),
-                    (method: VectorHelpers.MethodVectorValue.ForRaw, StringSegment stringSegmentAsBase64) => CoraxVectorItem.BuildSingleVector(builderParameters, fieldMetadata,
-                        GenerateEmbeddings.FromBase64Array(VectorOptions.Default, builderParameters.Allocator, stringSegmentAsBase64.ToString()), numberOfCandidates, minimumMatch, exact),
-                    (_, BlittableJsonReaderArray { Length: > 0 }) => throw new InvalidDataException("Cannot perform search on empty value."),
-                    _ => throw new InvalidQueryException(
-                        $"Unknown method in value ({vec.Method}. Parameter type: {methodParameter?.GetType().FullName}, Value: {methodParameter}")
-                };
-            }
-
-            embeddingsGenerationTaskIdentifier = vec.AiTaskName;
-            var vectorOptions = VectorHelpers.GetExplicitVectorOptions(builderParameters, fieldName, out indexField);
-            if (vectorOptions != null)
-            {
-                vectorOptions = new VectorOptions()
-                {
-                    DestinationEmbeddingType = vectorOptions.DestinationEmbeddingType,
-                    Dimensions = vectorOptions.Dimensions,
-                    SourceEmbeddingType = VectorEmbeddingType.Text,
-                    NumberOfCandidatesForIndexing = vectorOptions.NumberOfCandidatesForIndexing,
-                    NumberOfEdges = vectorOptions.NumberOfEdges
-                };
-            }
-
-            var vector = VectorHelpers.GetEmbeddingsForQueryParameter(builderParameters, valueTokenType, methodParameter, embeddingsGenerationTaskIdentifier, vectorOptions, fieldName);
-
-            if (vector.SingleVector != null)
-                return CoraxVectorItem.BuildSingleVector(builderParameters, fieldMetadata, vector.SingleVector.Value, numberOfCandidates, minimumMatch, exact);
-
-            return CoraxVectorItem.BuildMultiVector(builderParameters, fieldMetadata, vector.MultiVector, numberOfCandidates, minimumMatch, exact);
-        }
-
-        // Direct value (not a method call) — use pre-resolved value
-        var value = methodParameter;
-        var valueType = valueTokenType;
-
-        (VectorValue? SingleVector, VectorValue[] MultiVector) transformedEmbeddings = (null, null);
-        int numberOfDimensions;
-        if (VectorHelpers.TryRetrieveEmbeddingsGenerationTaskIdentifier(builderParameters, fieldName, out embeddingsGenerationTaskIdentifier))
-        {
-            var vectorOptions = VectorHelpers.GetExplicitVectorOptions(builderParameters, fieldName, out indexField);
-            transformedEmbeddings = VectorHelpers.GetEmbeddingsForQueryParameter(builderParameters, valueType, value, embeddingsGenerationTaskIdentifier, vectorOptions, fieldName);
-        }
-        else
-        {
-            VectorOptions vectorOptions = VectorHelpers.GetOptions(builderParameters, fieldName, out indexField);
-
-            if (builderParameters.Index.IndexFieldsPersistence.TryReadNumberOfDimensions(fieldName, out numberOfDimensions) == false)
-                return CoraxVectorItem.BuildEmpty(builderParameters); // no vector indexed
-            if (vectorOptions.SourceEmbeddingType is VectorEmbeddingType.Text)
-            {
-                transformedEmbeddings = VectorHelpers.GetVectorValueForTextualInput(builderParameters, vectorOptions, valueType, value);
-            }
-            else
-            {
-                switch (value)
-                {
-                    case string s:
-                        transformedEmbeddings.SingleVector = GenerateEmbeddings.FromBase64Array(vectorOptions, builderParameters.Allocator, s);
-                        break;
-                    case StringSegment stringSegment:
-                        transformedEmbeddings.SingleVector = GenerateEmbeddings.FromBase64Array(vectorOptions, builderParameters.Allocator, stringSegment.ToString());
-                        break;
-                    case BlittableJsonReaderObject bjro:
-                        transformedEmbeddings.SingleVector = VectorHelpers.GetVectorValueFromRavenVector(builderParameters, bjro, vectorOptions);
-                        break;
-                    case BlittableJsonReaderArray { Length: > 0 } bjra:
-                    {
-                        var isRavenVector = bjra[0] is BlittableJsonReaderObject;
-                        var isStringArray = bjra[0] is string or StringSegment or LazyStringValue;
-                        var isArray = bjra[0] is BlittableJsonReaderArray;
-
-                        if (isRavenVector == false && isStringArray == false && isArray == false)
-                        {
-                            transformedEmbeddings.SingleVector = VectorHelpers.GetVectorValueFromNumericalBlittableArray(builderParameters, bjra, vectorOptions);
-                        }
-                        else
-                        {
-                            var embeddings = new VectorValue[bjra.Length];
-                            for (int i = 0; i < bjra.Length; ++i)
-                            {
-                                if (isRavenVector)
-                                    embeddings[i] = VectorHelpers.GetVectorValueFromRavenVector(builderParameters, (BlittableJsonReaderObject)bjra[i], vectorOptions);
-                                else if (isStringArray)
-                                    embeddings[i] = GenerateEmbeddings.FromBase64Array(vectorOptions, builderParameters.Allocator, bjra[i].ToString());
-                                else
-                                    embeddings[i] = VectorHelpers.GetVectorValueFromNumericalBlittableArray(builderParameters, (BlittableJsonReaderArray)bjra[i],
-                                        vectorOptions);
-                            }
-
-                            transformedEmbeddings.MultiVector = embeddings;
-                        }
-
-                        break;
-                    }
-                    default:
-                        PortableExceptions.Throw<InvalidDataException>("We expected to get vector(s), however got: " + value.GetType().Name);
-                        break;
-                }
-            }
-        }
-
-        if (builderParameters.Index.IndexFieldsPersistence.TryReadNumberOfDimensions(fieldName, out numberOfDimensions) == false)
-            return CoraxVectorItem.BuildEmpty(builderParameters); // no vector indexed
-
-        if (transformedEmbeddings.SingleVector != null)
-        {
-            var singleVector = transformedEmbeddings.SingleVector.Value;
-
-            if (indexField != null)
-                AssertDimensions(singleVector);
-            return CoraxVectorItem.BuildSingleVector(builderParameters, fieldMetadata, singleVector, numberOfCandidates, minimumMatch, exact);
-        }
-
-        if (transformedEmbeddings.MultiVector != null)
-        {
-            var multiVector = transformedEmbeddings.MultiVector;
-
-            if (indexField != null)
-            {
-                foreach (var vector in multiVector)
-                    AssertDimensions(vector);
-            }
-
-            return CoraxVectorItem.BuildMultiVector(builderParameters, fieldMetadata, multiVector, numberOfCandidates, minimumMatch, exact);
-        }
-
-        throw new InvalidDataException("Expected to get single or multiple embeddings of VectorValue type but none was provided");
-
-        void AssertDimensions(in VectorValue vector)
-        {
-            if (numberOfDimensions != vector.Length)
-            {
-                using (vector)
-                    VectorHelpers.ThrowDifferentNumberOfDimensions(indexField, fieldName, vector, numberOfDimensions);
-            }
-        }
-    }
-
-    private static class VectorHelpers
-    {
-        public enum MethodVectorValue
-        {
-            ForDocument,
-            ForRaw,
-            EmbeddingText
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static bool TryRetrieveEmbeddingsGenerationTaskIdentifier(QueryBuilderParameters builderParameters, in string fieldName, out string embeddingsGenerationTaskIdentifier)
-        {
-            var existsInPersistence =
-                builderParameters.Index.IndexFieldsPersistence.TryReadEmbeddingsGenerationTaskIdentifier(fieldName, out embeddingsGenerationTaskIdentifier);
-
-            if (builderParameters.Metadata.IsDynamic == false)
-                return existsInPersistence;
-
-            if (((builderParameters.FieldsToFetch != null && builderParameters.FieldsToFetch.IndexFields.TryGetValue(fieldName, out var indexField)) || (builderParameters.Index.Definition.IndexFields.TryGetValue(fieldName, out indexField))) &&
-                indexField.Vector is AutoVectorOptions avo)
-            {
-                embeddingsGenerationTaskIdentifier = avo.EmbeddingsGenerationTaskIdentifier;
-                return string.IsNullOrEmpty(avo.EmbeddingsGenerationTaskIdentifier) == false;
-            }
-
-            embeddingsGenerationTaskIdentifier = null;
-            return false;
-        }
-
-        internal static (VectorValue? SingleVector, VectorValue[] MultiVector) GetVectorValueForTextualInput(QueryBuilderParameters parameters, VectorOptions vectorOptions, ValueTokenType valueType, object value)
-        {
-            if (valueType is ValueTokenType.String)
-                return (GenerateEmbeddings.FromText(parameters.Allocator, vectorOptions, value.ToString()), null);
-
-            if (valueType is not ValueTokenType.Parameter)
-                PortableExceptions.Throw<InvalidDataException>($"Cannot use vector.search() on a text field with a non-string value. Got {valueType}");
-
-            if (value is BlittableJsonReaderArray valueAsList)
-            {
-                var embeddings = new VectorValue[valueAsList.Length];
-                for (var i = 0; i < valueAsList.Length; ++i)
-                    embeddings[i] = GenerateEmbeddings.FromText(parameters.Allocator, vectorOptions, valueAsList[i].ToString());
-
-                return (null, embeddings);
-            }
-
-            PortableExceptions.Throw<InvalidDataException>($"Cannot use vector.search() on a text field with a non-string value(s). Got {valueType}");
-            return (null, null);
-        }
-
-        internal static VectorValue GetVectorValueFromRavenVector(QueryBuilderParameters parameters, BlittableJsonReaderObject json, VectorOptions vectorOptions)
-        {
-            var vectorObjectFound = json.TryGetMember(Sparrow.Global.Constants.Naming.VectorPropertyName, out var vectorObject);
-            PortableExceptions.ThrowIfNot<InvalidDataException>(vectorObjectFound, "Cannot find vector property in the object.");
-
-            var vectorReader = (BlittableJsonReaderVector)vectorObject;
-            return QueryBuilderHelper.GetVectorValueFromBlittableJsonVectorReader(parameters.Allocator, vectorOptions, vectorReader);
-        }
-
-        internal static VectorValue GetVectorValueFromNumericalBlittableArray(QueryBuilderParameters parameters, BlittableJsonReaderArray array, VectorOptions vectorOptions)
-        {
-            var bytesUsed = array.Length * (vectorOptions.SourceEmbeddingType is VectorEmbeddingType.Single ? sizeof(float) : 1);
-            var memScope = parameters.Allocator.Allocate(bytesUsed, out Memory<byte> mem);
-
-            // Hoist the per-element type switch out of the loop: SourceEmbeddingType is
-            // constant for the whole array, so dispatching once and then running a tight
-            // typed copy loop avoids repeating the branch (and the three ref captures) per
-            // element.
-            switch (vectorOptions.SourceEmbeddingType)
-            {
-                case VectorEmbeddingType.Single:
-                    CopyFloats(array, MemoryMarshal.Cast<byte, float>(mem.Span));
-                    break;
-                case VectorEmbeddingType.Int8:
-                    CopyInt8(array, MemoryMarshal.Cast<byte, sbyte>(mem.Span));
-                    break;
-                default:
-                    CopyBytes(array, mem.Span);
-                    break;
-            }
-
-            return GenerateEmbeddings.FromArray(parameters.Allocator, memScope, mem, vectorOptions, bytesUsed);
-
-            static void CopyFloats(BlittableJsonReaderArray src, Span<float> dst)
-            {
-                ref var dstRef = ref MemoryMarshal.GetReference(dst);
-                for (int i = 0; i < src.Length; ++i)
-                    Unsafe.Add(ref dstRef, i) = src.GetByIndex<float>(i);
-            }
-
-            static void CopyInt8(BlittableJsonReaderArray src, Span<sbyte> dst)
-            {
-                ref var dstRef = ref MemoryMarshal.GetReference(dst);
-                for (int i = 0; i < src.Length; ++i)
-                    Unsafe.Add(ref dstRef, i) = src.GetByIndex<sbyte>(i);
-            }
-
-            static void CopyBytes(BlittableJsonReaderArray src, Span<byte> dst)
-            {
-                ref var dstRef = ref MemoryMarshal.GetReference(dst);
-                for (int i = 0; i < src.Length; ++i)
-                    Unsafe.AddByteOffset(ref dstRef, i) = src.GetByIndex<byte>(i);
-            }
-        }
-
-        internal static VectorOptions GetExplicitVectorOptions(QueryBuilderParameters builderParameters, in string fieldName, out IndexField indexField)
-        {
-            if ((builderParameters.FieldsToFetch != null && builderParameters.FieldsToFetch.IndexFields.TryGetValue(fieldName, out indexField)) == false
-                && (builderParameters.Index.Definition.IndexFields.TryGetValue(fieldName, out indexField)) == false)
-                PortableExceptions.Throw<InvalidDataException>($"Cannot find `{fieldName}` field in the index.");
-
-            return indexField.Vector;
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static VectorOptions GetOptions(QueryBuilderParameters builderParameters, in string fieldName, out IndexField indexField)
-        {
-            if ((builderParameters.FieldsToFetch != null && builderParameters.FieldsToFetch.IndexFields.TryGetValue(fieldName, out indexField)) == false
-                && (builderParameters.Index.Definition.IndexFields.TryGetValue(fieldName, out indexField)) == false)
-                PortableExceptions.Throw<InvalidDataException>($"Cannot find `{fieldName}` field in the index.");
-
-            // VectorOptions can be null when a user does not specify the configuration.
-            // In such cases, we will choose the input depending on the value type (similar to how we handle it during indexing).
-            if (indexField.Vector != null)
-                return indexField.Vector;
-
-            builderParameters.Index.IndexFieldsPersistence.TryReadVectorSourceEmbeddingType(fieldName, out var vectorSourceEmbeddingType);
-
-            var defaultVectorOptions = vectorSourceEmbeddingType switch
-            {
-                VectorEmbeddingType.Single => VectorOptions.Default,
-                VectorEmbeddingType.Text => VectorOptions.DefaultText,
-                _ => throw new InvalidDataException(
-                    $"Unknown vector source embedding type: {vectorSourceEmbeddingType}. Implicit configuration support only single and text vector source embedding types.")
-            };
-
-            indexField.Vector = defaultVectorOptions;
-
-            return defaultVectorOptions;
-        }
-
-        internal static void ThrowDifferentNumberOfDimensions(in IndexField indexField, in string fieldName, in VectorValue transformedEmbedding,
-            in int numberOfDimensions)
-        {
-            var (storedDimensions, inputDimensions) = indexField.Vector.DestinationEmbeddingType switch
-            {
-                VectorEmbeddingType.Single => (numberOfDimensions / sizeof(float), transformedEmbedding.Length / sizeof(float)),
-                VectorEmbeddingType.Int8 => (numberOfDimensions - sizeof(float), transformedEmbedding.Length - sizeof(float)),
-                VectorEmbeddingType.Binary => (numberOfDimensions, transformedEmbedding.Length),
-                _ => throw new InvalidDataException($"Unexpected embedding type - {numberOfDimensions}.")
-            };
-
-            PortableExceptions.Throw<InvalidDataException>(
-                $"Vector field `{fieldName}` has {storedDimensions} dimensions, but the vector passed to vector.search() has {inputDimensions} dimensions.");
-        }
-
-        internal static (VectorValue? SingleVector, VectorValue[] MultiVector) GetEmbeddingsForQueryParameter(QueryBuilderParameters builderParameters, ValueTokenType valueType,
-            object value,
-            string embeddingsGenerationTaskIdentifier, VectorOptions vectorOptions, string fieldName)
-        {
-            var database = builderParameters.Index.DocumentDatabase;
-
-            var embeddingsTaskId = new EmbeddingsGenerationTaskIdentifier(embeddingsGenerationTaskIdentifier);
-
-            var embeddingsGenerator = database.EmbeddingsGeneratorQueries;
-
-            var sourceEmbeddingType = embeddingsGenerator.GetQuantizationOf(embeddingsTaskId);
-
-            // Quantized dynamic field indicates that the task generated embeddings with different quantization than requested in the index
-            // In this case we want to use quantization defined in dynamic field (which was set in CurrentIndexingScope.GetLoadVectorField)
-            VectorEmbeddingType destinationEmbeddingType;
-            if (builderParameters.Metadata.IsDynamic)
-            {
-                destinationEmbeddingType = sourceEmbeddingType is not VectorEmbeddingType.Single ? sourceEmbeddingType : vectorOptions!.DestinationEmbeddingType;
-            }
-            else
-            {
-                destinationEmbeddingType = vectorOptions?.DestinationEmbeddingType ?? sourceEmbeddingType;
-            }
-
-            ReadOnlyMemory<ReadOnlyMemory<byte>> embeddingValues;
-
-            switch (valueType)
-            {
-                case ValueTokenType.String:
-                    embeddingValues = embeddingsGenerator
-                        .GetEmbeddingsForQuery(builderParameters.DocumentsContext, embeddingsTaskId, value.ToString());
-                    break;
-                case ValueTokenType.Parameter:
-                {
-                    if (value is not BlittableJsonReaderArray bjra)
-                        throw new InvalidQueryException($"Expected array as parameter of vector.search({fieldName}) method, got '{value.GetType().FullName}' type instead.");
-
-                    var values = new string[bjra.Length];
-
-                    for (var i = 0; i < values.Length; i++)
-                        values[i] = bjra[i].ToString();
-
-                    embeddingValues = embeddingsGenerator
-                        .GetEmbeddingsForQuery(builderParameters.DocumentsContext, embeddingsTaskId, values);
-                    break;
-                }
-                default:
-                    throw new NotSupportedException($"Unexpected value type provided as parameter to vector.search({fieldName}) method. Got '{value.GetType().FullName}' type.");
-            }
-
-            var queryingVectorOption = new VectorOptions
-            {
-                SourceEmbeddingType = sourceEmbeddingType,
-                DestinationEmbeddingType = destinationEmbeddingType
-            };
-
-            if (embeddingValues.Length == 1)
-            {
-                var embeddingValue = embeddingValues.Span[0];
-
-                return (GenerateEmbeddings.FromArray(builderParameters.Allocator, embeddingValue.Span, queryingVectorOption), null);
-            }
-            else
-            {
-                var vectorValues = new VectorValue[embeddingValues.Length];
-
-                for (int i = 0; i < embeddingValues.Length; i++)
-                {
-                    var embeddingValue = embeddingValues.Span[i];
-
-                    vectorValues[i] = GenerateEmbeddings.FromArray(builderParameters.Allocator, embeddingValue.Span, queryingVectorOption);
-                }
-
-                return (null, vectorValues);
-            }
-        }
-    }
-
-    // ── Compound field merging (WHERE only, no ORDER BY) ──────────────────
-
     // ── Compound field exact match (no ORDER BY) ─────────────────────────
 
     public static bool TryCreateCompoundExactMatch(
@@ -3822,351 +3069,6 @@ internal static partial class QueryPlanBuilder
         return ds;
     }
 
-    // ── Sort seek hint ────────────────────────────────────────────────────
-
-    /// <summary>If the first clause is a range predicate on the same field as the first
-    /// ORDER BY field, set a seek hint on the CompiledQueryMatch so SortedIndexReader
-    /// can skip walking irrelevant tree terms.</summary>
-    public static void TrySetSortSeekHint(CompiledQueryMatch match,
-        QueryExecution exec, OrderMetadata[] orderByFields)
-    {
-        if (orderByFields == null || orderByFields.Length == 0)
-            return;
-
-        var execs = exec.Executions;
-        if (execs == null || execs.Length == 0)
-            return;
-
-        // Only consider the first ORDER BY field
-        var sortField = orderByFields[0].Field.FieldName;
-
-        // Find a range clause on the same field (scan all clauses, not just first — the sort-eligible
-        // clause may not be the cheapest and thus not clause[0]).
-        for (int i = 0; i < execs.Length; i++)
-        {
-            var clause = execs[i].Clause;
-            var sortExec = execs[i];
-
-            if (clause.FieldName != sortField.ToString())
-                continue;
-
-            // Range clauses on the sort field — supports long, double, and string
-            if (clause.ClauseType is not (ClauseType.GreaterThan or ClauseType.GreaterThanOrEqual
-                or ClauseType.LessThan or ClauseType.LessThanOrEqual or ClauseType.Between))
-                continue;
-
-            var packed = sortExec.PackedParamValue;
-            if (packed.IsNone)
-                continue;
-
-            // For ascending order: seek to the lower bound (GT/GTE value)
-            // For descending order: seek to the upper bound (LT/LTE value)
-            bool ascending = orderByFields[0].Ascending;
-            object seekValue = null;
-
-            if (ascending && clause.ClauseType is ClauseType.GreaterThan or ClauseType.GreaterThanOrEqual)
-            {
-                seekValue = packed.ValueType switch
-                {
-                    PackedParam.TypeLong => exec.LongValues[packed.Param1],
-                    PackedParam.TypeDouble => exec.DoubleValues[packed.Param1],
-                    PackedParam.TypeString => exec.StringValues[packed.Param1],
-                    _ => null
-                };
-            }
-            else if (ascending == false && clause.ClauseType is ClauseType.LessThan or ClauseType.LessThanOrEqual)
-            {
-                seekValue = packed.ValueType switch
-                {
-                    PackedParam.TypeLong => exec.LongValues[packed.Param1],
-                    PackedParam.TypeDouble => exec.DoubleValues[packed.Param1],
-                    PackedParam.TypeString => exec.StringValues[packed.Param1],
-                    _ => null
-                };
-            }
-            else if (clause.ClauseType == ClauseType.Between)
-            {
-                // Between: seek to the lower bound for ASC, upper bound for DESC
-                int idx = ascending ? packed.Param1 : packed.Param2;
-                seekValue = packed.ValueType switch
-                {
-                    PackedParam.TypeLong => exec.LongValues[idx],
-                    PackedParam.TypeDouble => exec.DoubleValues[idx],
-                    PackedParam.TypeString => exec.StringValues[idx],
-                    _ => null
-                };
-            }
-
-            if (seekValue != null)
-            {
-                match.SortHint = new SortHint(clause.FieldName, seekValue);
-                return;
-            }
-        }
-    }
-
-    public static OrderMetadata[] GetSortMetadata(QueryBuilderParameters builderParameters, out bool hasEmpty)
-    {
-        hasEmpty = false;
-        var query = builderParameters.Query;
-        var index = builderParameters.Index;
-        var getSpatialField = builderParameters.Factories?.GetSpatialFieldFactory;
-        var indexMapping = builderParameters.IndexFieldsMapping;
-        var queryMapping = builderParameters.FieldsToFetch;
-        var allocator = builderParameters.Allocator;
-        if (query.PageSize == 0) // no need to sort when counting only
-        {
-            return null;
-        }
-
-        var orderByFields = query.Metadata.OrderBy;
-
-        if (orderByFields == null)
-        {
-            if (builderParameters.HasBoost && (
-                    index.Configuration.OrderByScoreAutomaticallyWhenBoostingIsInvolved
-                    || index.Configuration.CoraxVectorSearchOrderByScoreAutomatically))
-            {
-                if (builderParameters.Metadata.HasVectorSearch == false)
-                    builderParameters.IndexReadOperation?.AssertCanOrderByScoreAutomaticallyWhenBoostingOrVectorSearchIsInvolved();
-
-                return [new OrderMetadata(true, MatchCompareFieldType.Score)];
-            }
-
-            return null;
-        }
-
-        if (orderByFields.Length == 0)
-            return null;
-
-        if (orderByFields.Length > MaxSortFields)
-            throw new InvalidOperationException($"Corax does not support ordering by more than {MaxSortFields} properties.");
-
-        // OrderMetadata contains a managed IPoint reference field, so stackalloc is not
-        // possible here (only unmanaged structs may be stack-allocated in C#).
-        int sortIndex = 0;
-        var sortArray = new OrderMetadata[MaxSortFields];
-
-        foreach (var field in orderByFields)
-        {
-            var nullsSortMode = (field.NullsOrdering, field.Ascending) switch
-            {
-                (NullsOrderingType.First, Ascending: true) => NullsSortMode.NullsSmallest,
-                (NullsOrderingType.First, Ascending: false) => NullsSortMode.NullsLargest,
-                (NullsOrderingType.Last, Ascending: true) => NullsSortMode.NullsLargest,
-                (NullsOrderingType.Last, Ascending: false) => NullsSortMode.NullsSmallest,
-                _ => (NullsSortMode?)null
-            };
-
-            if (field.OrderingType == OrderByFieldType.Random)
-            {
-                var seed = field.Arguments is { Length: > 0 } ? (int)Hashing.XXHash32.CalculateRaw(field.Arguments[0].NameOrValue) : Random.Shared.Next();
-                sortArray[sortIndex++] = new OrderMetadata(seed);
-                continue;
-            }
-
-            if (field.OrderingType == OrderByFieldType.Score)
-            {
-                // EntryComparerByScore.Compare is intentionally inverted (returns y.CompareTo(x)),
-                // so ascending=true -> highest scores first (the default "most relevant first" search engine order).
-                // ascending=false -> Descending<EntryComparerByScore> -> lowest scores first.
-                sortArray[sortIndex++] = new OrderMetadata(true, MatchCompareFieldType.Score, field.Ascending);
-
-                continue;
-            }
-
-            var fieldMetadata = QueryBuilderHelper.GetFieldIdForOrderBy(allocator, field.Name, index, builderParameters.HasDynamics,
-                builderParameters.DynamicFields, indexMapping, queryMapping, false);
-
-            bool fieldIsEmpty = builderParameters.IndexSearcher.GetTermAmountInField(fieldMetadata) == 0;
-            if (fieldIsEmpty)
-            {
-                if (builderParameters.IndexReadOperation.IsSharded == false)
-                    continue;
-                hasEmpty = true;
-            }
-
-            if (field.OrderingType == OrderByFieldType.Distance)
-            {
-                var spatialField = getSpatialField(field.Name);
-
-                int lastArgument;
-                IPoint point;
-                switch (field.Method)
-                {
-                    case MethodType.Spatial_Circle:
-                        var cLatitude = field.Arguments[1].GetDouble(query.QueryParameters);
-                        var cLongitude = field.Arguments[2].GetDouble(query.QueryParameters);
-                        lastArgument = 2;
-                        point = spatialField.ReadPoint(cLatitude, cLongitude).Center;
-                        break;
-                    case MethodType.Spatial_Wkt:
-                        var wkt = field.Arguments[0].GetString(query.QueryParameters);
-                        SpatialUnits? spatialUnits = null;
-                        lastArgument = 1;
-                        if (field.Arguments.Length > 1)
-                        {
-                            spatialUnits = Enum.Parse<SpatialUnits>(field.Arguments[1].GetString(query.QueryParameters), ignoreCase: true);
-                            lastArgument = 2;
-                        }
-
-                        point = spatialField.ReadShape(wkt, spatialUnits).Center;
-                        break;
-                    case MethodType.Spatial_Point:
-                        var pLatitude = field.Arguments[0].GetDouble(query.QueryParameters);
-                        var pLongitude = field.Arguments[1].GetDouble(query.QueryParameters);
-                        lastArgument = 2;
-                        point = spatialField.ReadPoint(pLatitude, pLongitude).Center;
-                        break;
-                    default:
-                        throw new ArgumentOutOfRangeException();
-                }
-
-                var roundTo = field.Arguments.Length > lastArgument
-                    ? field.Arguments[lastArgument].GetDouble(query.QueryParameters)
-                    : 0D;
-
-                sortArray[sortIndex++] = new OrderMetadata(fieldMetadata, field.Ascending, MatchCompareFieldType.Spatial, point, roundTo,
-                    spatialField.Units is SpatialUnits.Kilometers
-                        ? global::Corax.Utils.Spatial.SpatialUnits.Kilometers
-                        : global::Corax.Utils.Spatial.SpatialUnits.Miles, fieldIsEmpty, nullsSortMode);
-                continue;
-            }
-
-            var orderingType = field.OrderingType;
-            if (orderingType is OrderByFieldType.Implicit && index.Configuration.OrderByTicksAutomaticallyWhenDatesAreInvolved && index.IndexFieldsPersistence.HasTimeValues(field.Name.Value))
-                orderingType = OrderByFieldType.Long;
-
-            var metadataField = QueryBuilderHelper.GetFieldIdForOrderBy(allocator, field.Name.Value, index, builderParameters.HasDynamics,
-                builderParameters.DynamicFields,
-                indexMapping, queryMapping, false);
-            // Dynamic CreateField fields: no IndexFieldsMapping entry, FieldId == DynamicField (-2).
-            // Such fields are written per-document only when the index function emits CreateField;
-            // docs that don't emit the field have NO entry (not even a NonExisting marker) in the
-            // field's tree, so StreamAndIntersect (which walks tree + null/nonExisting lists) would
-            // silently drop them. Route through ExtractAndSort instead — see SortingMatch.Fill.
-            bool mayHaveMissingEntries = metadataField.FieldId == Constants.IndexWriter.DynamicField;
-            OrderMetadata? temporaryOrder = null;
-            switch (orderingType)
-            {
-                case OrderByFieldType.Custom:
-                    throw new NotSupportedInCoraxException($"{nameof(Corax)} doesn't support Custom OrderBy.");
-                case OrderByFieldType.AlphaNumeric:
-                    sortArray[sortIndex++] = new OrderMetadata(metadataField, field.Ascending, MatchCompareFieldType.Alphanumeric, fieldIsEmpty, nullsSortMode, mayHaveMissingEntries);
-                    continue;
-                case OrderByFieldType.Long:
-                    temporaryOrder = new OrderMetadata(metadataField, field.Ascending, MatchCompareFieldType.Integer, fieldIsEmpty, nullsSortMode, mayHaveMissingEntries);
-                    break;
-                case OrderByFieldType.Double:
-                    temporaryOrder = new OrderMetadata(metadataField, field.Ascending, MatchCompareFieldType.Floating, fieldIsEmpty, nullsSortMode, mayHaveMissingEntries);
-                    break;
-            }
-
-            sortArray[sortIndex++] = temporaryOrder ?? new OrderMetadata(metadataField, field.Ascending, MatchCompareFieldType.Sequence, fieldIsEmpty, nullsSortMode, mayHaveMissingEntries);
-        }
-
-        return sortArray[0..sortIndex];
-    }
-
-    public static IQueryMatch OrderBy(QueryBuilderParameters builderParameters, IQueryMatch match, in OrderMetadata[] orderMetadataSource, bool hasEmptySortingMatches)
-    {
-        RuntimeHelpers.EnsureSufficientExecutionStack();
-        var indexSearcher = builderParameters.IndexSearcher;
-        var take = builderParameters.Take;
-        OrderMetadata[] orderMetadata = null;
-        if (hasEmptySortingMatches == false)
-            orderMetadata = orderMetadataSource;
-        else
-        {
-            var currentIdx = 0;
-            foreach (var orderMetadataItem in orderMetadataSource)
-            {
-                if (orderMetadataItem.FieldHasNoTerms)
-                    continue;
-
-                orderMetadata ??= new OrderMetadata[orderMetadataSource.Length];
-                orderMetadata[currentIdx++] = orderMetadataItem;
-            }
-
-            orderMetadata = currentIdx == 0 ? [] : orderMetadata![..currentIdx];
-        }
-
-        switch (orderMetadata.Length)
-        {
-            case 0:
-                return match;
-            case 1:
-                return indexSearcher.OrderBy(match, orderMetadata[0], builderParameters.Index.Configuration.NullsSortMode, take, builderParameters.Token);
-            default:
-                return indexSearcher.OrderBy(match, orderMetadata, builderParameters.Index.Configuration.NullsSortMode, take, builderParameters.Token);
-        }
-    }
-
-    /// <summary>Apply ORDER BY from plan metadata when a full <see cref="QueryBuilderParameters"/> is not
-    /// available (e.g., direct tests). Handles <c>ORDER BY score()</c> only — callers that need
-    /// field / spatial / alphanumeric sorts must use the full
-    /// <see cref="OrderBy(QueryBuilderParameters,IQueryMatch,in OrderMetadata[],bool)"/> overload.</summary>
-    public static IQueryMatch ApplyScoreOrdering(QueryPlanBuilder.PlanParameters planParams, IQueryMatch match, long take, CancellationToken token = default)
-    {
-        OrderByField[] orderByFields = planParams.Metadata.OrderBy;
-        if (orderByFields == null || orderByFields.Length == 0)
-            return match;
-
-        var indexSearcher = planParams.IndexSearcher;
-        int takeInt = take > int.MaxValue ? Constants.IndexSearcher.TakeAll : (int)take;
-
-        for (int i = 0; i < orderByFields.Length; i++)
-        {
-            if (orderByFields[i].OrderingType == OrderByFieldType.Score)
-            {
-                var meta = new OrderMetadata(true, MatchCompareFieldType.Score, orderByFields[i].Ascending);
-                return indexSearcher.OrderBy(match, meta, NullsSortMode.NullsLargest, take: takeInt, token: token);
-            }
-        }
-
-        return match;
-    }
-
-    // ── Search helpers (used by ResolveClause for Search clause type) ────
-
-    /// <summary>
-    /// Replaces the search analyzer with an appropriate wildcard analyzer.
-    /// LuceneAnalyzerAdapter wrapping KeywordAnalyzer has IsExactAnalyzer=false
-    /// (because LuceneAnalyzerAdapter passes NoTransformers), so the generic
-    /// CreateWildcardAnalyzer in the Legacy path would incorrectly lowercase the term.
-    /// This matches the old CoraxQueryBuilder.ReplaceAnalyzerForWildcardQueries logic.
-    /// </summary>
-    private static FieldMetadata ReplaceAnalyzerForWildcardQueries(
-        FieldMetadata searchMeta,
-        ResolutionContext walkerCtx)
-    {
-        var result = searchMeta;
-        var indexFieldsMapping = walkerCtx.BuilderParams.IndexFieldsMapping;
-
-        if (searchMeta.IsDynamic && indexFieldsMapping != null)
-            result = searchMeta.ChangeAnalyzer(searchMeta.Mode, indexFieldsMapping.SearchAnalyzer(searchMeta.FieldName.ToString()));
-
-        if (searchMeta.Analyzer is Lucene.LuceneAnalyzerAdapter laa && indexFieldsMapping != null)
-        {
-            global::Corax.Analyzers.Analyzer replacementAnalyzer = laa.Analyzer switch
-            {
-                global::Lucene.Net.Analysis.KeywordAnalyzer => indexFieldsMapping.ExactAnalyzer(searchMeta.FieldName.ToString()),
-                Lucene.Analyzers.RavenStandardAnalyzer
-                    or Lucene.Analyzers.NGramAnalyzer => indexFieldsMapping.DefaultAnalyzer,
-                global::Lucene.Net.Analysis.Standard.StandardAnalyzer when laa.Analyzer.GetType() == typeof(global::Lucene.Net.Analysis.Standard.StandardAnalyzer)
-                    => indexFieldsMapping.DefaultAnalyzer,
-                Lucene.Analyzers.LowerCaseKeywordAnalyzer
-                    or Lucene.Analyzers.Collation.CollationAnalyzer => indexFieldsMapping.DefaultAnalyzer,
-                _ => null
-            };
-
-            if (replacementAnalyzer != null)
-                result = searchMeta.ChangeAnalyzer(global::Corax.FieldIndexingMode.Search, replacementAnalyzer);
-        }
-
-        return result;
-    }
-
     /// <summary>Create the appropriate DirectScan match based on whether residual predicates exist.</summary>
     private static DirectScanMatchBase BuildDirectScan(
         IndexSearcher searcher, IQueryMatch drivingMatch,
@@ -4952,52 +3854,6 @@ internal static partial class QueryPlanBuilder
         // so we iterate it directly without materializing into a bitmap first.
         return [new PlanOp { Kind = PlanOpKind.DirectIterate, ParamIndex = 0 }];
     }
-
-    private static void AttachPostFilterPhases(QueryExecution exec, bool allNegated,
-        ClauseInfo[] spatialClauses, ClauseExecution[] spatialExecs,
-        ClauseInfo[] vectorClauses, ClauseExecution[] vectorExecs)
-    {
-        if (spatialClauses == null && vectorClauses == null)
-        {
-            return;
-        }
-
-        // Extend Executions array to include spatial/vector post-filter clauses.
-        int extraCount = (spatialClauses?.Length ?? 0) + (vectorClauses?.Length ?? 0);
-        var execs = exec.Executions ??= [];
-        int execIdx = execs.Length;
-        Array.Resize(ref execs, execs.Length + extraCount);
-
-        int matchIndex = CountMatchSlots(execs, exec.IsAllEntries, allNegated);
-
-        if (spatialClauses != null)
-        {
-            exec.SpatialFilters = new SpatialFilterOp[spatialClauses.Length];
-            for (int i = 0; i < spatialClauses.Length; i++)
-            {
-                var spatialExec = spatialExecs?[i] ?? new ClauseExecution(spatialClauses[i]);
-                execs[execIdx++] = spatialExec;
-                exec.SpatialFilters[i] = new SpatialFilterOp { MatchIndex = matchIndex++, Clause = spatialClauses[i], Exec = spatialExec };
-            }
-        }
-
-        if (vectorClauses != null)
-        {
-            exec.VectorSelects = new VectorSearchOp[vectorClauses.Length];
-            for (int i = 0; i < vectorClauses.Length; i++)
-            {
-                var vectorExec = vectorExecs?[i] ?? new ClauseExecution(vectorClauses[i]);
-                execs[execIdx++] = vectorExec;
-                exec.VectorSelects[i] = new VectorSearchOp
-                {
-                    Clause = vectorClauses[i], Exec = vectorExec
-                };
-            }
-        }
-
-        exec.Executions = execs;
-    }
-
 
     internal static int CountMatchSlots(ClauseExecution[] executions, bool isAllEntries, bool allNegated)
     {
